routes reorganized.

+++ src/handlers/projectHandler.go
@@ -4,10 +4,13 @@ import (
 	"artanis/src/configs"
 	"artanis/src/models"
 	"artanis/src/models/base"
+	"artanis/src/models/clients"
+	"artanis/src/models/enums"
 	"artanis/src/models/requests"
 	"artanis/src/models/responses"
 	"artanis/src/repositories/projectRepository"
 	"github.com/gofiber/fiber/v2"
+	"github.com/google/uuid"
 )
 
 type ProjectHandler struct {
@@ -21,13 +24,20 @@ func NewProjectHandler(db *projectRepository.ProjectRepository, cfg *configs.Con
 
 func (h *ProjectHandler) Register(c *fiber.Ctx) error {
 	var projectRequest requests.RegisterProject
+	tokenUser := c.Context().UserValue("user").(*clients.User)
 	if err := c.BodyParser(&projectRequest); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(base.Error{Message: err.Error()})
 	}
 
+	if tokenUser.OrganizationRole == enums.OrganizationUser {
+		return c.Status(fiber.StatusForbidden).JSON(base.Error{Message: "Not enough credentials to create a project"})
+	}
+
 	project := models.Project{
-		Name:        projectRequest.Name,
-		Description: projectRequest.Description,
+		Id:             uuid.New().String(),
+		Name:           projectRequest.Name,
+		Description:    projectRequest.Description,
+		OrganizationId: tokenUser.OrganizationId,
 	}
 
 	if err := h.db.RegisterProject(project); err != nil {
@@ -41,11 +51,11 @@ func (h *ProjectHandler) Register(c *fiber.Ctx) error {
 }
 
 func (h *ProjectHandler) Paginate(c *fiber.Ctx) error {
+	user := c.Context().UserValue("user").(*clients.User)
 	limit := c.QueryInt("limit")
 	offset := c.QueryInt("offset")
-	organizationId, _ := c.Locals("organizationId").(string)
 
-	projects, err := h.db.PaginateProjects(organizationId, limit, offset)
+	projects, err := h.db.PaginateProjects(user.OrganizationId, limit, offset)
 
 	if err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(base.Error{Message: "Failed to register the project"})
@@ -59,11 +69,15 @@ func (h *ProjectHandler) Paginate(c *fiber.Ctx) error {
 }
 
 func (h *ProjectHandler) Update(c *fiber.Ctx) error {
+	tokenUser := c.Context().UserValue("user").(*clients.User)
+	if tokenUser.OrganizationRole == enums.OrganizationUser {
+		return c.Status(fiber.StatusForbidden).JSON(base.Error{Message: "Not enough credentials to create a project"})
+	}
+
 	var projectRequest requests.UpdateProject
 	if err := c.BodyParser(&projectRequest); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(base.Error{Message: err.Error()})
 	}
-
 	if err := h.db.UpdateProject(projectRequest.Id, projectRequest.Name, projectRequest.Description); err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(base.Error{Message: "Failed to update the project"})
 	}
@@ -75,6 +89,11 @@ func (h *ProjectHandler) Update(c *fiber.Ctx) error {
 }
 
 func (h *ProjectHandler) Delete(c *fiber.Ctx) error {
+	tokenUser := c.Context().UserValue("user").(*clients.User)
+	if tokenUser.OrganizationRole == enums.OrganizationUser {
+		return c.Status(fiber.StatusForbidden).JSON(base.Error{Message: "Not enough credentials to create a project"})
+	}
+
 	projectId := c.Params("id")
 
 	if err := h.db.DeleteProject(projectId); err != nil {
@@ -88,8 +107,8 @@ func (h *ProjectHandler) Delete(c *fiber.Ctx) error {
 }
 
 func (h *ProjectHandler) GetDashboardCounts(c *fiber.Ctx) error {
-	organizationId := c.Query("organizationId")
-	counts, err := h.db.GetDashboardCounts(organizationId)
+	user := c.Context().UserValue("user").(*clients.User)
+	counts, err := h.db.GetDashboardCounts(user.OrganizationId)
 	if err != nil {
 		return err
 	}
@@ -108,6 +127,7 @@ func mapPaginateResponse(projects []models.Project) []responses.ProjectResponse 
 			Id:          project.Id,
 			Name:        project.Name,
 			Description: project.Description,
+			CreatedAt:   project.CreatedAt,
 		})
 	}
 	return projectsResponse
